refactor(main): pure utilities を src/app/diff_viewer/util.rs に切り出す (#224 step 2)
DIFF_APP_STATE thread_local や Slint UI ハンドルに依存しない純粋関数を
別 module へ移動。

切り出し対象:
- send_mode_label
- current_hhmmss
- resolve_line_number
- short_sha
- excerpt

既存の resolve_line_number / short_sha 系 unit test も util.rs に移し、
さらに excerpt の boundary case (blank / truncate / first paragraph) を
新規追加。

差分:
- src/main.rs: 1726 → 1650 行 (-76)
- src/app/diff_viewer/util.rs: 121 行 (新規 + 8 unit test)
- 残置: refresh_* / show_toast / schedule_toast_auto_dismiss / append_history
  などは DIFF_APP_STATE thread_local / ACTIVE_DIFF_WINDOW を使うため、別 step
  で extract する

検証: cargo clippy / cargo test (143 passed; +3 新規)

Refs: #224
