+++ src/app/diff_viewer/mod.rs
@@ -4,3 +4,4 @@
 //! 状態 / ヘルパは段階的にここへ移動する (#224 split)。
 
 pub(crate) mod state;
+pub(crate) mod util;
